After crazy refactoring, the first simple quad exports again correctly. Used different wording to describe the parts of a mesh, using "elements" for the point/normal/tangent vectors, and "components" for the x,y,z,w coordinates. Now more testing is needed on the other scenes, but committing this first.

+++ src/SceneTypes.h
@@ -62,8 +62,8 @@ namespace Semantic
 		return static_cast<Kind>(s);
 	}
 
-	// Get the number of components per semantic 
-	inline size_t dimension(const Kind s)
+	// Get the number of float components per semantic 
+	inline int dimension(const Kind s)
 	{
 		switch (s)
 		{
